requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pillow
numpy
opencv-python
torch
torchvision
ultralytics
open_clip_torch